backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-google-genai
python-dotenv
python-multipart
requests
pydantic